[Refactor] major JIT kernel clean up for dsv4 (#25884)
