+++ python/sglang/jit_kernel/dsv4/__init__.py
@@ -1,4 +1,32 @@
-from .compress import *
+from .attn import (
+    fused_store_cache,
+    get_paged_mqa_logits_metadata,
+    triton_create_paged_compress_data,
+)
+from .compress import (
+    CompressorDecodePlan,
+    CompressorPrefillPlan,
+    compress_forward,
+    compress_norm_rope_store,
+)
+from .compress_old import fused_norm_rope_inplace
+from .elementwise import (
+    fused_k_norm_rope_flashmla,
+    fused_q_indexer_rope_hadamard_quant,
+    fused_q_norm_rope,
+    fused_rope_inplace,
+)
+from .gemm import linear_bf16_fp32
+from .hisparse import hisparse_offload_to_host
+from .moe import (
+    hash_topk,
+    mask_topk_ids,
+    mega_moe_pre_dispatch,
+    silu_and_mul_clamp,
+    silu_and_mul_contig_post_quant,
+    silu_and_mul_masked_post_quant,
+)
+from .topk import plan_topk_v2, topk_transform_512, topk_transform_512_v2
 from .utils import make_name
 
 __all__ = [
@@ -6,5 +34,24 @@ __all__ = [
     "CompressorPrefillPlan",
     "compress_forward",
     "compress_norm_rope_store",
+    "fused_norm_rope_inplace",
+    "fused_store_cache",
+    "fused_rope_inplace",
+    "fused_q_norm_rope",
+    "fused_q_indexer_rope_hadamard_quant",
+    "fused_k_norm_rope_flashmla",
     "make_name",
+    "linear_bf16_fp32",
+    "hisparse_offload_to_host",
+    "get_paged_mqa_logits_metadata",
+    "triton_create_paged_compress_data",
+    "topk_transform_512",
+    "topk_transform_512_v2",
+    "plan_topk_v2",
+    "hash_topk",
+    "mega_moe_pre_dispatch",
+    "mask_topk_ids",
+    "silu_and_mul_clamp",
+    "silu_and_mul_masked_post_quant",
+    "silu_and_mul_contig_post_quant",
 ]
