+++ python/sglang/srt/layers/moe/moe_runner/triton_utils/fused_moe.py
@@ -569,7 +569,7 @@ def _fused_moe_kernel_sequence(
 
             if not filter_expert:
                 if swiglu_limit_for_silu_and_mul_clamp is not None:
-                    from sglang.jit_kernel.deepseek_v4 import silu_and_mul_clamp
+                    from sglang.jit_kernel.dsv4 import silu_and_mul_clamp
 
                     silu_and_mul_clamp(
                         intermediate_cache1.view(-1, N),
